+++ CHANGES.md
@@ -5,6 +5,7 @@
 * In protocol v3, clamp out-of-range timestamps rather than erroring. [Fixes #14](https://github.com/ewxrjk/sftpserver/issues/14).
 * Subsecond timestamps are now returned on POSIX platforms, where supported. [Fixes #12](https://github.com/ewxrjk/sftpserver/issues/12).
 * In protocol v3, correctly report large file sizes in the human-readable part of the output.  [Fixes #15](https://github.com/ewxrjk/sftpserver/issues/15).
+* The SFTP client tracks the working directory properly. [Fixes #11](https://github.com/ewxrjk/sftpserver/issues/11).
 
 ## Changes in version 2